feat(cubejs-cli): Move helpers to TypeScript
