+++ packages/cubejs-cli/src/index.js
@@ -24,4 +24,4 @@ if (major < 8) {
   process.exit(1);
 }
 
-require('./cubejsCli');
+require('./cli');
